@mcollina 2019-02-04
Would this make it compatitble with fluent-schema?