@Eomm 2019-02-04
> Would this make it compatitble with fluent-schema?

Yes, generally with this fix, the `schema.response.200` settings in fastify will accept the `$ref = '#customId ` attribute without failing (already tried to run this fix in fastify and works, will send a PR with this specific test 👍)